# Conflicts:
#	src/main/java/com/example/final_project/controller/AccountController.java
#	src/main/java/com/example/final_project/dto/AccountDTO.java
#	src/main/java/com/example/final_project/entities/Account.java
#	src/main/java/com/example/final_project/repository/AccountRepository.java
#	src/main/java/com/example/final_project/service/AccountService.java
#	src/main/resources/application.properties

+++ src/main/java/com/example/final_project/dto/TransactionDTO.java
@@ -2,21 +2,19 @@ package com.example.final_project.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-
 import java.time.LocalDateTime;
+
 @Getter
 @Setter
 public class TransactionDTO {
     private Long id;
-    private Long fromAccountNumber;
-    private Long toAccountNumber;
+    private Integer fromAccountNumber;
+    private Integer toAccountNumber;
     private Double amount;
     private String transactionType;
     private LocalDateTime timestamp;
-    // getters and setters
 
-
-    public TransactionDTO(Long id, Long fromAccountNumber, Long toAccountNumber, Double amount, String transactionType, LocalDateTime timestamp) {
+    public TransactionDTO(Long id, Integer fromAccountNumber, Integer toAccountNumber, Double amount, String transactionType, LocalDateTime timestamp) {
         this.id = id;
         this.fromAccountNumber = fromAccountNumber;
         this.toAccountNumber = toAccountNumber;
@@ -24,6 +22,7 @@ public class TransactionDTO {
         this.transactionType = transactionType;
         this.timestamp = timestamp;
     }
-    public TransactionDTO() {
-    }
+
+    public TransactionDTO() {}
 }
+
